+++ eras/conway/impl/src/Cardano/Ledger/Conway.hs
@@ -21,7 +21,6 @@ import Cardano.Ledger.Alonzo.Language (Language (..))
 import Cardano.Ledger.Alonzo.TxInfo (EraPlutusContext, ExtendedUTxO (..))
 import Cardano.Ledger.Babbage.Tx (babbageTxScripts)
 import Cardano.Ledger.Babbage.TxBody ()
-import Cardano.Ledger.Binary (sizedValue)
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Era (ConwayEra)
 import Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
@@ -35,12 +34,6 @@ import Cardano.Ledger.Conway.UTxO ()
 import Cardano.Ledger.Crypto (Crypto, StandardCrypto)
 import Cardano.Ledger.Keys (DSignable, Hash)
 import qualified Cardano.Ledger.Shelley.API as API
-import Cardano.Ledger.UTxO (UTxO (..))
-import Data.Foldable (toList)
-import qualified Data.Map.Strict as Map
-import Data.Maybe.Strict
-import qualified Data.Set as Set
-import Lens.Micro
 
 type Conway = ConwayEra StandardCrypto
 
@@ -78,9 +71,3 @@ instance
   where
   txInfo = conwayTxInfo
   txscripts = babbageTxScripts
-  getAllowedSupplimentalDataHashes txBody (UTxO utxo) =
-    Set.fromList [dh | txOut <- outs, SJust dh <- [txOut ^. dataHashTxOutL]]
-    where
-      newOuts = map sizedValue $ toList $ txBody ^. allSizedOutputsTxBodyF
-      referencedOuts = Map.elems $ Map.restrictKeys utxo (txBody ^. referenceInputsTxBodyL)
-      outs = newOuts <> referencedOuts